=General *Fixed a error that appears on certain setups for the packet_ver_flag -setting. Guess its set as a signed int. So that means only 2 slots -are left for newer clients?
*Updated the item_db perl script.



=SQL
*Updated the item_db, item_db_3ceam, and item_db_custom files to
-set the equip_locations as int(11) as its needed to properly store
-the equip location values for shadow equips. This is needed to
-properly show the shadow equips equip locations in flux control
-panel. User's will need to rebuild their item databases as the
-old int(4) setting causes equip locations values for shadow equips
-to be set to the old max value of 65535.



=NPC
*Updated the Global_Functions weapon and armor values.



=Documents
*Updated the item_db document.

git-svn-id: svn://svn.code.sf.net/p/v1-3ceam/code/trunk@810 cad27aaa-dce3-4a30-a00a-e4fd67c11881

+++ rewrite/sql-files/item_db_3ceam.sql
@@ -18,7 +18,7 @@ CREATE TABLE `item_db_3ceam` (
   `equip_jobs` int(12) unsigned default NULL,
   `equip_upper` tinyint(8) unsigned default NULL,
   `equip_genders` tinyint(2) unsigned default NULL,
-  `equip_locations` smallint(4) unsigned default NULL,
+  `equip_locations` int(11) unsigned default NULL,
   `weapon_level` tinyint(2) unsigned default NULL,
   `equip_level` tinyint(3) unsigned default NULL,
   `refineable` tinyint(1) unsigned default NULL,
